I init new component <LaunchPage /> and <DashboardPage />
