+++ components/DashboardPage/index.ts
@@ -1,0 +1,3 @@
+import DashboardPage from "./DashboardPage";
+
+export default DashboardPage
