+++ components/LaunchPage/index.ts
@@ -1,0 +1,3 @@
+import LaunchPage from "./LaunchPage";
+
+export default LaunchPage
